Remove calcSkip and formatPage from MongoDriver
