+++ src/LearningObjectSearch/index.ts
@@ -8,7 +8,7 @@ import { ElasticSearchLearningObjectDatastore } from './drivers';
 import { UserGateway } from './interfaces/UserGateway';
 import { UserServiceGateway } from './gateways/UserServiceGateway';
 import { UserLearningObjectDatastore } from './interfaces/UserLearningObjectDatastore';
-import { MongoDBLearningObjectDatastore } from './drivers/LearningObjectDatastore/MongoDB/MongoDBLearningObjectDatastore/MongoDBLearningObjectDatastore';
+import { MongoDBLearningObjectDatastore } from './drivers/UserLearningObjectDatastore/MongoDB/MongoDBLearningObjectDatastore/MongoDBLearningObjectDatastore';
 
 @expressServiceModule({
   expressRouter: ExpressHttpAdapter.buildRouter(),
