+++ src/app.ts
@@ -9,7 +9,7 @@ import {
   DataStore,
 } from './shared/interfaces/interfaces';
 import { LibraryDriver } from './drivers/LibraryDriver';
-import { MongoConnector } from './shared/Mongo/MongoConnector';
+import { MongoConnector } from './shared/MongoDB/MongoConnector';
 import { LearningObjectAdapter } from './LearningObjects/adapters/LearningObjectAdapter';
 import { LearningObjectSearch } from './LearningObjectSearch';
 import { HierarchyAdapter } from './LearningObjects/Hierarchy/HierarchyAdapter';
